Tpetra: Resolve transform code to remove incorrect fence
After discussion we determined withLocalAccess needs a fix
which will fix the incorrect sync state after a transform to host.
Such fix is also expected to make CUDA_LAUNCH_BLOCKING=0 work
since the sync will call the required fences.

I'm removing the fence because even though it makes the test pass
with CUDA_LAUNCH_BLOCKING=0, it's not the right solution and stacked
on the withLocalAccess issue.