MueLu: adding thresholding of the interpolation values in GeometricInterpolationPFactory, see issue #6681
This is a first cut that performs reasonably well, the idea is to look at interpolation coefficient generated for each fine point.
If the coefficients are smaller than a given tolerance (1e-6), they are dropped.
This results in much sparser coarse grid as well as smaller P and R matrices.
If we want to keep this idea around we should rewritte the structured aggregation factory to obtain the correct graph directly instead of recomputing it in the prolongator factory.