PyTrilinos: Clean up Isorropia module warnings
Isorropia::Epetra has several constructors and methods that overshadow
each other in the python implementation because an argument in one is
a pointer, and the argument in the other is an RCP of the same type.
I explicitly ignore the raw pointer versions to prevent the warnings
when running SWIG on Isorropia.Epetra.i.