Tpetra: Fix CUDA build error in pad_crs_arrays
@trilinos/tpetra

That was supposed to be a host parallel loop, but I was using a
device RangePolicy.  This commit fixes that.

I also removed the DeviceType template parameter from CrsPadding.
CrsPadding was not using that template parameter, and the class is
an implementation detail anyway, so we can always put it back if
we need it.  I don't like to give the impression that a class is
using Kokkos and/or running on device when it's not.