Redo configure and build for CUDA-aware test run build (CDOFA-72)
This results in configure info, cache vars, etc. going up in the build so that
developers can see that info.  Since this is doing a rebuild and the version
of Trilinos does not change, the rebuild should not build any targets and
should finish almost immediately.

This was requested by @kddevin in #6796 as it was confusing that you can't see
configure info for these special "cuda-aware-mpi" builds otherwise.