Framework: Fixing PR test driver python script tests
The mock testing in the PR test driver python scripts was working
on specific machines with certain CPU counts, but it didn't hanlde
systems outside a specific set (i.e., my mac laptop).

This update fixes this so that we are correctly mocking a set
number of 'cpus' that the testing scripts will see so that the
tests should work on more systems.