Framework: Enhance failure messaging in PR test driver
The PR test driver needs either psutil to be installed
or /proc/meminfo should exist (i.e., linux).  First it tries
to get system memory information via psutil, which is the preferred
method. If psutil isn't installed and thus fails, it will attempt
to get the information by parsing /proc/meminfo which is generally
available on *nix systems.

If both of these methods fail (i.e., you're on a mac without psutil
installed) then we catch that IOError and raise a new IOError with
a more descriptive message than the default.