Make work with /bin/sh and not just bash on 'vortex'
This was requested by EMPIRE that uses Python subprocess and runs /bin/sh and
not bash to source this somehow.  I don't know that it is a good idea to not
explictly call a bash script and then source this inside of a new bash process
but this is what Scot Swan requested and says works on 'vortex' with the
EMPIRE python Trilinos build script (see
https://cee-gitlab.sandia.gov/EM-Plasma/BuildScripts/issues/230) so we will go
with it.  I don't see the harm in making this change but I am concerned about
signing up to support sourcing this script in a non-bash raw /bin/sh process
since it was never the plan to support that use case.  (So perhaps we should
have used the extension *.bash instead of *.sh?)