Tpetra::CrsGraph: Add nodeMinNumRowEntries_
@trilinos/tpetra

Add private instance variable nodeMinNumRowEntries_ to
Tpetra::CrsGraph.  This will be useful in Tpetra::CrsMatrix for
deciding which sparse matrix-vector multiply kernel to use.  We don't
have to expose it in CrsGraph's public interface, since CrsMatrix is
already a friend of CrsGraph.

Also, make nodeMinNumRowEntries_ and nodeMaxNumRowEntries_ both
local_ordinal_type instead of size_t.  This is the appropriate type
for these two quantities.  Doing so means that adding
nodeMinNumRowEntries should not make sizeof(CrsGraph) increase.