Tpetra: Add test for LocalCuSparseCrsMatrixOperator::apply; it passes
This test is one level of abstraction higher (more abstract) than the
low-level CuSparseMatrix test I added just before this commit.
LocalCuSparseCrsMatrixOperator needs to create a copy of the input
matrix's row offsets array.  This is because cuSPARSE expects the row
offsets to have the same integer type as the local column indices, but
this is not true with KokkosSparse::CrsMatrix by default.