Remove unused script, reuse bash func, and update doc (CDOFA-102)
We just use raw bsub and lrun commands to run on the compute node (that is
simple and easy to document).  We don't seem to need a complex driver script
like we need for SLURM systems.

However, it is nice having a named function for getting the allocated compute
node name and SSHing to it.