Tpetra: CUDA_LAUNCH_BLOCKING unset fixes
These are additional fixes identified on a machine
with an Nvidia GTX960 card. These errors did not exhibit
on white.

This loop for example will always bus error on the GTX960 system
but on white does not crash to the extent tested. A fence
before the std::cout is fine on both machines as expected.

  const int N = 100000;
  Kokkos::View<double*> input("input", N);
  Kokkos::View<double*> output("output", N);
  Kokkos::parallel_for(Kokkos::RangePolicy<>(0,N), KOKKOS_LAMBDA(const size_t i) {
    output(i) = input(i) + 1.0;
  });
  std::cout << "Check can read: " << input(0) << std::endl;