MueLu: fixing a couple of issues in pairwise aggregation.
Now the tests are passing for multiple aggregation iterations
in serial and in parallel. Also removed some debugging print
statement and added very basic statistics.
Final step will be to add an xml parameter for the number of
iterations and to create an input deck and a test using the
scaling driver.