Stokhos:  Fix MueLu ETI includes
The first include in the auto-generated MueLu ETI files is
MueLu_ConfigDefs.hpp to pull in any configure-time preprocessor macros.
Unfortunately this file also includes Kokkos headers that need to come
after the Stokhos includes (so that the Stokhos overloads are in scope
at the right time), causing compiler errors.  This commit replaces the
include with MueLu_config.hpp, so that the macros are still present
without the Kokkos includes.