MueLu: region restriction operator factory
This factory makes assumptions about the problem to allow for
faster algorithms to compute the restriction operator.
Implemented all the geometric cases independently to
fill the restriction matrix. This should allow for
good performance, hopefully faster than what smoothed
aggregation does on a Laplace 3D problem.
Just multiply the fine level nullspace by R to construct
the coase nullpsace, injection could be performed instead.
Also adding the logic in the unit test to check that
the coarse nullspace can be pulled out from the level
and inspected.
Adding some logic so that coordinates are now propagated
as the nullspace.
Adding a regression test that runs both regionRFactory
and GeometricInterpolationPFactory on the same mesh.
It the compares the residual history of both and determines
if there is any difference between them (tol=1.0e-12)