ATDM: van1-tx2: Address problem of 'van1-tx2' matching KOKKOS_ARCH=TX2 (#4063, CDOFA-27)
After the "upgrade" of the ATDM Trilinos parsing system to parse out the
KOKKOS_ARCH in either upper or lower case, now the 'tx2' in 'van1-tx2' matches
the KOKKOS_ARCH 'TX2' (lower case 'tx2').  This broke the running of
atdm/van1-tx2/environment.sh.  This broke the nightly ATDM Trilinos 'van1-tx2'
builds and installs of Trilinos from 'stria' today, testing day 2020-04-21.

I updated the error message to show the KOKKOS_ARCH and then I updated the
logic to accept the KOKKOS_ARCH of 'TX2'.  That is a bit of a hack but it gets
us around this problem (and is harmless on this system).