Clean up the NetCDF library settings.
For SEMS, the pnetcdf library is in the SEMS_NETCDF_ROOT directory and
SEM_PNETCDF_ROOT is not defined (or PNETCDF_ROOT).  In all scripts checked
whether PNETCDF_ROOT was defined and if not, set it to value of NETCDF_ROOT

In cmake config files make sure root of libpnetcdf.{a|so} is PNETCDF_ROOT and
make sure PNETCDF_ROOT was set to some value.