SEACAS, Percept: don't try to install disabled exe
Executables created with TRIBITS_ADD_EXECUTABLE have "categories" for
what kind of tests they are used for (e.g. BASIC, PERFORMANCE). The
default is BASIC, which means that for a build that does not include
that category, the executable won't exist. Now checking that the
executable target exists before trying to install it.