Panzer: set virtual cells to have appropriate normals and rotation matrices.
The rule is supposed to be that virtual cells have always have face ordinal 0 being the one that abuts the real cell, and that these have normals that are the negative of the normal on the real cell at the same point.  This commit makes the panzer Workset adhere to this rule.