Curl library dependency not related to MPI
The curl library dependency is independent of parallel (MPI) setting. It is a consequence of enabling DAP2 or DAP4 support in NetCDF. It is correctly detected if the FindNetCDF.cmake module is called, but that does not seem to be called for the Trilinos builds that I have done; probably due to the environment variables set before cmake is called?.  

The sems builds do have DAP enabled and therefore require the curl library.

Here we enable the curl dependency if the FindNetCDF.cmake module is not called. This works on all the pull-request builds and several ATDM HPC builds that I have tried.