ifpack2: Reduce memory usage of Ifpack2::IlukGraph::initialize().
In particular for the case where there are a small number of significantly
more dense rows. Rather than using the max number of entries from any
row as the heuristic for every row instead use the per-row values.
This reduces the memory use of a particular test case from ~1.2GB
down to ~350MB (which is consistent with the equivalent Aztec solver
memory use).