Kokkos: Fix span calculation of View for corner case
This bug was affecting only subviews taken in a way that
for a LayoutLeft view the first subview argument was a pair of length
zero, or for a LayoutRight the last argument was a pair of length zero
but all other dimensions are larger than 0. This is a bug which
has likely always existed in Kokkos (I checked back to 2.7, but don't
see a reason why this wouldn't have been here before).