Amesos2: Clean up loading for Kokkos adapter
No longer loads via Tpetra and converts to Kokkos.
Just load directly via Kokkos calls.

Only loads data for the root rank and empty for others.
This mimics the behavior you would get for Tpetra and
something like SuperLU if the test runs on more than 1 rank.

For Tpetra we load the matrix distributed but then
the adapter collects it all to the root rank and the others are empty.

For kokkos just load everything initially on the root rank
and leave the others empty,

This fixes the crash that would occur if you tried to
run SuperLU on multiple ranks.

Also resolves refactor setup was not getting randomized for any of the tests..