MueLu: region matvec clean-up a bit the setup
The matvec setup is now implementated and setup and apply are tested
in the fastMatVec unit-test.
Tested by comparing the fast MatVec results with the old "slow"
implementation of the MatVec.
When you're not interested in the timing results, you might want to just
not print them to the screen to have a little less cluttered screen
output.
Modify V-cycle to allow for the fast MatVec implementation.
The coarse multivectors are computed using the prolongator graph,
thus no communication is required in that step.
Next a coarse version of interfaceGIDsMV is needed to call SetupMatVec
on the coarse levels.
Finishing the computation of coarse MatVec data in MakeCoarseLevelMaps.
Adding interface to createRegionHierarchy such that it outputs MatVec data
on each level.
Finally store the MatVec data on the smoother parameter list.
Working fast MatVec implemented on all levels.