CONFIG: Update NetCDF and HDF5 TPL version (#7197)
* CONFIG: Update NetCDF and HDF5 TPL version

* DO NOT MERGE See if this fixes builds

* CONFIG: Update sems netcdf module extension for new versions

* NETCDF: Remove BoostLib dependency

* NETCDF: See if this triggers use of curl

* NETCDF: Better way of specifying library list

* HDF5 -- See what HDF5 variables are set to...

* Undo some changes that were probably not needed

* NETCDF: Try older versions on gcc-4.8.4 build

* NETCDF: Try older versions on gcc-4.8.4 build

* NETCDF: Remove incorrect boost dependency

* Eliminate debug print

* Clean up the NetCDF library settings.

For SEMS, the pnetcdf library is in the SEMS_NETCDF_ROOT directory and
SEM_PNETCDF_ROOT is not defined (or PNETCDF_ROOT).  In all scripts checked
whether PNETCDF_ROOT was defined and if not, set it to value of NETCDF_ROOT

In cmake config files make sure root of libpnetcdf.{a|so} is PNETCDF_ROOT and
make sure PNETCDF_ROOT was set to some value.

* TPLs updated on test machine

Use HDF5-1.10.6 and NetCDF-4.7.3 now.

* Eliminate curl library dependency for cuda build

The SEMS installations of NetCDF included the DAP setting which requires -lcurl;

Installations on other systems did not include DAP, so don't need curl.  This is detected in the FindNetCDF.cmake, but that is currently not called for most builds...

* Mutrino NetCDF does not require curl library

* Ride NetCDF library does not require curl library

* shiller NetCDF does not require curl library

* van1-tx2 NetCDF library does not require curl library

* waterman NetCDF library does not require curl library

* Remove debug variable prints

* Update the non-unit test file also

After sems was updated on test machine, I updated the unit test file on 4.8.4 build, but not the other file that also needed updating...

* Curl library dependency not related to MPI

The curl library dependency is independent of parallel (MPI) setting. It is a consequence of enabling DAP2 or DAP4 support in NetCDF. It is correctly detected if the FindNetCDF.cmake module is called, but that does not seem to be called for the Trilinos builds that I have done; probably due to the environment variables set before cmake is called?.  

The sems builds do have DAP enabled and therefore require the curl library.

Here we enable the curl dependency if the FindNetCDF.cmake module is not called. This works on all the pull-request builds and several ATDM HPC builds that I have tried.

Co-authored-by: Greg Sjaardema <gdsjaar@sandia.gov>