ATDM: sems-rhel6: Add missing boost module load (#7275)
The change the sems-netcdf/4.4.1/exo_parallel module to the
sems-netcdf/4.7.3/parallel module in commit 09e765d merged in PR #7179
resulted in the sems-boost/1.59.0/base module no longer being implicilty
loaded.  This was causing a configure-time failure where boost could not be
found.

This commit adds a explicit load of sems-boost/1.59.0/base.  This allows the
configure of Trilinos to work (but I have not tested a full set of builds).