Set default for PANZER_FADTYPE at all points
I was seeing intermittent failures to build
with PANZER_FADTYPE undefined. The value was
set using TRIBITS_ADD_OPTION_AND DEFINE which
expects a bool and was passed a string. For
whatever reason, cmake is inconsistent about
how the passed string is interpreted. Since
there is no provision for this not being set
I just went to a direct SET statement.