Fixes errors in OrthoManagerTest and enables Epetra testing
The OrthoManagerTest had several typos in it, which were not found
because all tests that used it were essentially disabled.  The
OrthoManagerTest has been fixed and all TSQR-specific tests have
been placed behind an ifdef.  The Epetra interface test for the
OrthoManager has been enabled and the dependency on TSQR has been
removed due to the change in the tester.