Adds missing norm calculation in singular vector detector for block orthogonalization
In both the ICGS and IMGS orthogonalization the calculation of the
newDot values were missing in the blkOrthoSing method.  This method
is only used in block orthogonalization, when the operator for the
linear system is ill-conditioned such that the basis becomes degenerate.
So only BlockGmres and BlockCG could have encountered this issue with
block sizes larger than 1.

A test has been added to the Epetra interface to ensure that this
issue doesn't happen again.