Amesos2: Remove duplication of KokkosSparse_sptrsv_superlu.hpp file
This is not a final solution.
It calls 'using namespace SLU' in a header which is not desirable.

However it does remove the duplicate code and the above using is
only used for builds with TRSV enabled. So it is probably a better
short term solution until we decide how to make kernels and Amesos2
have a uniform namespace convention.