TrilinosCouplings:  Fix uninitialized variable use in many examples.
All of these examples have the same pattern of using an array of scalars
without initializing them:

Scalar flux[3] = {0.0, 0.0, 0.0};

...

for (int i = 0; i < 3; ++i)
  for (int j = 0; j < 3; ++j)
    flux[i] += material[i][j]*grad_u[j];

In this case Scalar = SFad<double,3>.  This worked with the old design,
because SFad always zero-initialized its derivative array, but currently
the new design does not.  It is unclear if the new design should be
changed to maintain consistency, but I decided to go ahead and change
these examples anyway, because I don't think any reasonable person would
expect the code to work correctly as it was written.  For most scalar
types, including builtin ones, the code wouldn't have worked correctly.