IOSS: Fix setting of exodiff binary in tests
A previous commit tried to improve the setting of the path to the
exodiff binary, but mistakenly added a method that only worked on
a preexisting build.  On a clean build, it would result
in an empty symbol which resulted in an empty program name following
the CMND keyword.

This fixes that problem.