Add test for ShyLU_NodeHTS and list all packages for CDash (#6697, #7229)
I added an explict test for the subpackage ShyLU_NodeHTS to make sure it will
get tested if there are any changes (see #7229).

I also changed the list of toplevel packages as CDash subproject to list all
of the PR packages.  This is because you can't just list the changed SE
packages, you have to list all of the top-level packages that have any enabled
subpackages downstream.  That type of logic will require doing a full TriBITS
dependency analysis in CMake code calling the macro
TRIBITS_ADJUST_AND_PRINT_PACKAGE_DEPENDENCIES().  That will require some
refactoring that I don't have the time to do now.

But note that at least the list of toplevel packages will always be correct as
there are changes to the Trilinos/PackagesList.cmake file.  That is better
than using a static list that never gets updated.