Fixes calls to normalize zero vectors in ICGS/IMGS
The logic that checks if the norm of a vector is zero before normalizing
it needs to be inserted.  This conditional existed in DGKS, which is why
this did not come up as a problem before the change to making ICGS[2] the
default.

The OrthoManager test has been amended to include a check for this
functionality, so that it does not come up again for customers.