Sacado:  Add configure option to force SFad to init in default constructor
The new design SFad does not directly initialize its data in the default
constructor, but instead uses the default compiler generated
implementation.  This speeds SFad up but has proved to be problematic in
some codes that were assuming SFad always initialized (as was the case
in the old design).  This adds a configure option,
Sacado_SFAD_INIT_DEFAULT_CONSTRUCTOR, to recover the old behavior.  It
defaults to off.

Also add KOKKOS_DEFAULTED_FUNCTION in several places for Cuda 9 and
lower.