Sacado:  Add Sacado::safe_sqrt(x) function for handling x == 0
The derivative of sqrt(x) is not defined for x == 0.  Computing the
derivative with x == 0 happens frequently in application codes, and the
usual advice is to return 0 in this case, to avoid the NaN.  It would be
convenient to have a function that does this for you automatically, so I
added a safe_sqrt() function that does just that.  It includes a default
implementation that just calls sqrt(), has overloads for all of the
primary Fad types, and is SIMD-safe.