Adds example/test for multi-matrix solves using pseudo-block GMRES
The pseudo-block solvers have the ability to solve a set of linear
systems simultaneously.  This has to be defined by the linear problem,
which will manage the active set of linear problems being solved and
ensure the correct operator is applied.

Some of the methods in the Belos::LinearProblem class needed to be
virtual to enable this.  The example/test in the MultiMatrixSolve
directory of the Epetra interface shows how to create a new
LinearProblem class and solve an unpreconditioned set of linear systems.