Tpetra: added constructors for FECrsGraph
The new ctors accept an additional Map, representing the domain map
for the owned+shared graph.

Other ctors are implemented in terms of the new ones, and they pass
the domain map (for the owned graph) as input for the owned+shared
domain map.

If the input owned+shared domain map is null, these ctors use
ownedPlusSharedRowMap, which ensures that the latter is locally
fitted to the column map of the owned+shared graph.

Reason for this change: by passing a separate domain map for both owned and
owned+shared graphs, we ensure that both graphs have a domain map that
is locally fitted to the column map.

This is useful during FE assembly: inside FE kernels, one may have access only
to the ownedPlusShared map coming from some sort of discretization/connectivity
object, and not the col map of the matrix (which may or may not be
locally available in the assembly kernels). When converting GIDs of FE
dofs to LID, to be used to modify the local matrix, one needs the
owned+shared GIDs to appear first (and in the same order) as they
appear inside the FECrs matrix col map.