Framework: Adding in file=sys.stdout to print()'s
Some of the print() statements should have retained the `file=sys.stdout`
arguments since they were supposed to print to stderr.  I've re-added
them as placeholders.

Later we should look at this and determine if sending the error messages
to stderr is the right thing to do or if we should just keep them
sending to stdout.  Things to consider before changing to STDERR:
1. Does the autotester slurp in stderr output when it's capturing logs
   that it adds to PR's? If not then we should not send this to stderr
   since that might potentially lose error messages that could be
   useful.
2. Update python script testing to account for different outputs.