MueLu: use fast MatVec for regionMG prolongation
Prolongation is a region-local operation, that does not require a
summation of interface values. Hence, an option is added to ApplyMatVec
to disable summation of interface values. For now, this option defaults
to 'true' to maintain backwards compatibility.

Next step: remove the default value to force the user to explicitly
his/her wishes.

Part of #7175.