Framework: PRDriverTest.py - stringify envvars
There are some places where we're setting os environment vars but
using commands like `os.environ[key] = [value]` but a `TypeError`
will be thrown if the value is not a string.

This can happen easily if we set something in `environMap` to be
an integer vs. a string.  For example:

```python
"OMP_NUM_THREADS": 2
```
vs.
```python
"OMP_NUM_THREADS": "2"
```

The latter is what has been done but the first one can be done easily.

The easy solution to this is to try and force the values to strings
at assignment time, `os.environ[key] = str(value)`