Tempus:  Add setObserver(), initialize() funcs for issue #7357
This adds the setObserver() and initialize() functions to the adjoint
sensitivity integrator as requested in issue #7357.  These just call the
corresponding functions on the forward and adjoint integrators.  Note it
is is unclear how observes should work with adjoint sensitivities, so I
did not add the observer to the adjoint integrator.  That can easily be
changed if required.