Tpetra: WIP: fix KokkosKernels types to use ETI
Use "size_t" explicitly as the "offset" type for StaticCrsGraph
and KokkosSparse::CrsMatrix. Otherwise, the type is the default
size_type for the memory space which is "unsigned" for Cuda,
meaning that Tpetra's calls to KokkosKernels were not using the ETI'd
versions. Verified that Tpetra is now calling the ETI versions.

Still to do: add UVM and complex specializations for cuSPARSE SPMV.
That way merge path can be used by Tpetra in practice.