Tpetra: optionally disable ownership check in FECrsGraph
The user is now allowed to disable the check that ensures that for each
owned element, there is at least one owned node. This check is helpful
to keep the column map optimally sized, but for some application it might
be difficult to rearrange the mesh ownership, to the point that it might
be easier to accept a small performance hit during mat-vec product (due
to import of not needed gids).

Fixes #7697.