Tpetra: improved ownership check in FECrsGraph
Instead of just a flat bool signaling whether all is good, we now count
the number of col GIDs that are not connected to any owned row, and we
print this information (both local and global count) in the exception
message.

Also, changed the name of the option, since the old name was stating
a different check from the one we are actually checking.