FEI patch for astra from Mike glass.
From his email -

I don't know if any other ASC apps are using the FEI package in Trilinos but we (Sierra) have come across a compiler issue on stria related to this package. We're using the ARMHPC 20.1 toolset. The following is our work around. Can this get incorporated into the Trilinos repo?