MueLu: adding PlaneDetectionFactory stub for ice sheet project
From this point the number of planes or the number of nodes per
plane need to be passed by the user. The orientation: `plane` or
`line` also need to be specified. With these two parameters a
map of the planes can easily be computed to create a partitioner
for Ifpack2.