Turn off a kokkos-kernels test that runs over 600s
This was missed as timouts are set only at the dashboard level
in tribits. That meant that this did not show on the command
line. It consistently runs about 1080s in debug.

a) ATDM turned this same test off for debug back in Feb 2020 at commit e399ec64
b) kokkos-kernels looks to have tried to address this in Feb 2019 at commit c4d81605

As such I expect this is well understood by the kokkos-kernels team?