Adds logic for MPI builds where MPI has not been initialized
Jennifer Loe has been using an MPI build of Trilinos, but performing single
node execution.  In this situation, MPI is not initialized, so a different
communicator needs to be used.  Now the class checks if it's an MPI build
and if MPI has been initialized before using an MPI communicator.  Otherwise,
it uses a serial communicator.