Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This date call is designed to work with all variants of date.

Also use UTC to be independent of timezone.

This PR was done while working on reproducible builds for openSUSE.