ATDM: Remove checks for running on login and launch nodes (ATDV-387)
They updated 'vortex' to change the launch node from 'vortex59' to 'vortex5'
and that caused all of the Trilinos tests run with 'trilinos_jsrun' to fail
with a bogus error message.  Removing these checks makes these scripts more
robust to changes on 'vortex'.  Yes, it removes some checks to protect
developers using this system but if developers follow the instructions then
they will not have a problem.

We should have never of been checking for specific names of nodes on 'vortex'.
That was just makes the scripts more difficult to maintain (which is the last
thing we need at this point).