MueLu: move regionMG source code to new directory
This code started as a highly experimental research effort in
packages/muelu/research/mmayr/composite_to_region/. We consider this
still research and experimental, yet have rewritten most of it, so we
move it to another directory, packages/muelu/research/regionMG/src/, in
order to separate the files, that are well-maintained and still used
from other files in the original directory. This hopefully helps
unfamiliar developer to more easily understand the code base.

Part of #7824.