Tpetra: optionally disable ownership check in FECrsGraph (#7773)
* Tpetra: optionally disable ownership check in FECrsGraph

The user is now allowed to disable the check that ensures that for each
owned element, there is at least one owned node. This check is helpful
to keep the column map optimally sized, but for some application it might
be difficult to rearrange the mesh ownership, to the point that it might
be easier to accept a small performance hit during mat-vec product (due
to import of not needed gids).

Fixes #7697.

* Tpetra: make ownership check in FECrsGraph MPI aware

In parallel, perform a reduction of the success flag, to avoid leaving some ranks hanging.

* Tpetra: overridden impl of getValidParameters for FECrsGraph

* Tpetra: improved ownership check in FECrsGraph

Instead of just a flat bool signaling whether all is good, we now count
the number of col GIDs that are not connected to any owned row, and we
print this information (both local and global count) in the exception
message.

Also, changed the name of the option, since the old name was stating
a different check from the one we are actually checking.