Panzer: Fix Panzer_Filtered_GlobalIndexer for CUDA_LAUNCH_BLOCKING=0
putScalar will by default write to device.
replaceGlobalValue will always write to host.

So we can sync_host like this after putScalar to make sure the device write
is synched for UVM to host before replaceGlobalValue.

Then since replaceGlobalValue does not call modify_host, I think we need
to call this afterwards. That won't be necessary for UVM, but later if
device is CudaSpace, then skipping that call means the export will be confused
about where the data is and try to pull it from device.

An alternative is to call modify_host before the putScalar calls.
Then the putScalar will write to host and we can skip the modify_host
calls after the replaceGlobalValue. This might be confusing as I'm not
sure calling modify_host to force putScalar to host is a normal pattern.