Piro: introducing 'Constant Mass Matrix' option to tell Piro whether or not the mass matrix is changing in time.
Previously the code assumed this was 'true', which is not the
case for general dynamic problems.  The code was hence doing the
wrong thing for problems where the mass matrix was not constant in time.

The new option allows the user to tell the code if the mass matrix is
constant in time or not for the various time-steppers callable from Piro.

The option is only relevant for explicit time-stepper.

The default value of the 'Constant Mass Matrix' option is now false,
meaning that the code will by default assume the mass matrix is dependent on
time, so it will recompute the mass matrix at every time step.  This can
be turned off by setting the option to true - then the same mass matrix
will be used throughout the calculation as in the first time step.