Added apply function to vector class and updated affected code.
When the dot product associated with a vector space X is defined using
a non-identity matrix S, i.e.,

  x.dot(y) = x'*S*y,

we view S*y as being in the dual space of X.  The inner product on the
dual space is defined using inv(S).  In many locations, we need to
compute products of vectors in X with vectors in its dual space.  For
example, if g is the gradient, g is in the dual spaces of X.  Let s
be a step computed in X, then to check descent, we must apply g to x.
The previous version of ROL would compute this as

  g.dot(s.dual()),

which requires the application of S to s and then the application of inv(S)
to compute the dot product, i.e.,

  g.dot(s.dual()) = g'*inv(S)*(S*s).

The apply function allows us to avoid the redundant mat-vec computations.
In particular, the apply function for dual vectors acts on primal vectors
and similarly for dual vectors, e.g.,

  g.apply(s) = g'*s.

We view this as applying the linear functional g to the vector s.