Framework: PullRequestLinuxDriverMerge.py py2-3 compatibility fix
The calls to `subprocess.check_output()` in the Merge script do not
return the same thing in python2 and python3.

In python2 the output is a string object, but in python3 it is a bytes
object. This causes problems in Python3 if you do something like:

```python
output = subprocess.check_output(['ls'])

if 'foo.txt' in output:
    do_something()
```

In this code, Python2 will work just fine in the `if` statmement because
you're comparing strings. However, in Python3 this statement will fail
because the comparison is between a `bytes` object and a `str` object.

Adding the `output = output.decode('utf-8')` convert the output object
into a utf-8 string, which is compatible.
Python2 converts it to a 'unicode' type but it's still compatible with
a string comparison.