Stokhos,Tpetra:  Fix DistObject kernels for Stokhos using SFINAE.
The partial specializations of the DistObject kernels in Stokhos were
not actually being called by Tpetra because of a mismatch in the
template parameters.  To fix this, I changed the default implementations
in Tpetra to have an extra "Enabled" template parameter that Stokhos can
specialize using SFINAE, instead of trying to match the View parameters.
This required several other changes, including making the combine
operators non-template classes with templated operator(), and allowing
Stokhos' custom execution policy to support tags.  This was tested on
both Serial and Cuda builds.