MueLu: Ifpack2: implement lumped diagonal scaling
This implements a matrix diagonal scaling suggested by Ray Tuminaro.
The default is still to use the inverse of the point diagonal.
This commit adds an option to use as the diagonal the row sums
of absolute values:

   d_ii = sum(abs(A(ii,:)))

The reciprocal of this "lumped" diagonal  is used in the power method
used to estimate the maximum eigenvalue of D \ A and in the Chebyshev smoother.