Tpetra:  Add CombineMode ADD_ASSIGN for correct handling of apply with beta != 0 and non-null exporter (#7898)
* tpetra:  test reproducing bug in #7758

* tpetra:  added Epetra versions of the tests, to compare Epetra to Tpetra behavior
The behavior is the same.
These tests will need to be removed if we decide to fix only Tpetra -- not Epetra.

* Epetra: fixes for #7758, in which Export with CombineMode == Add produces
inconsistent results for source maps that are not supersets of target maps
Added zero-out of target vector to prevent inclusion of existing target values

* tpetra:  moved Epetra tests to epetra/test/MultiVector/Bug7758.cpp

* tpetra:  fixes for #7758, in which Export with CombineMode == ADD produces
inconsistent results for source maps that are not supersets of target maps
Added zero-out of target vector to prevent inclusion of existing target values.
Added CombineMode argument to copyAndPermute

* tpetra:  new argument (CombineMode) to copyAndPermute

* epetra: fixed test output messages

* tpetra:  changes to remove compiler warnings, handle complex

* Tpetra:  fixing one more compiler warning

* Tpetra:  Adding unit test for #7745
Also adding extra test for #7758

* tpetra:  not yet working, but need to archive it

* tpetra:  comment fix

* tpetra:  removed reset to zero in copyAndPermute; we don't want that #7758

* epetra:  removing initialize to zero; we don't really want it #7758

* epetra:  fix comments

* tpetra:  changed "ACCUMULATE" to "ADD_ASSIGN"
Added correctness tests to Bug7745 Multivector test using ADD_ASSIGN

* tpetra:  adding tests for bug 7745

* tpetra:  added square matrix option

* tpetra:  test transpose and refactor

* Tpetra:  changed transpose test

* tpetra:  added cyclic map tests

* tpetra:  changed ADD to ADD_ASSIGN in apply with non-null exporter
Fixed correctness error in batchApply (beta multiplied in too often)
added test for batchApply with non-null exporter

* tpetra:  added transpose test

* stokhos:  changes to match new prototypes for #7898

* tpetra:  removed unused variable warnings

* tpetra:  cuda complained about Kokkos::parallel_for lambdas in a
protected function.  Turning the lambda into a functor solved the
compilation problem.

* stokhos: Making op() calls match those for unpack

* Stokhos:  Add tag argument to PermuteArrayMultiColumn DistObject kernels

* Stokhos,Tpetra:  Fix DistObject kernels for Stokhos using SFINAE.

The partial specializations of the DistObject kernels in Stokhos were
not actually being called by Tpetra because of a mismatch in the
template parameters.  To fix this, I changed the default implementations
in Tpetra to have an extra "Enabled" template parameter that Stokhos can
specialize using SFINAE, instead of trying to match the View parameters.
This required several other changes, including making the combine
operators non-template classes with templated operator(), and allowing
Stokhos' custom execution policy to support tags.  This was tested on
both Serial and Cuda builds.

* tpetra:  added range policy so function executes in right place with
SerialNode

* tpetra: #7898 added comment regarding SFINAE template parameter that
makes Stokhos' specializations happy

Co-authored-by: Eric Phipps <etphipp@sandia.gov>