PR Driver Script Overhaul - Style Cleanup
Cleaning up some of the banners in the new PR Driver script
to make it look better when the console log is pasted into a
Pull Request on Github.

The issue has been that Github treats it like markdown so
banners like this:

===================
= My Nifty Banner
===================

Confuse Markdown into thinking that whatever is above each of the
`=============` lines is a Heading-1.

A simple fix to prepend a `+` to each line breaks this pattern.
I added a plus as a postfix as well for symmetry ;)