Some fixes in the test program.
Not running in parallel (,yet). Please note: on the finest level we provide
the striding information in the maps for the primal and dual dofs. No need to
define them in the xml file any more (in fact, the xml files would overwrite
this information on all levels, which is not what we need here. Maybe one would
need to change that behavior in Xpetra...)

Please note, that in the "mySubBlockAFactory1" i commented out the "Range map: Striding info"
and "Domain map: Striding info" settings (they are implicitly provided by the
program on the finest level.