MueLu: Allow both Tpetra and Eptra versions of duplicate tests (#8110)
Before, just the first duplicate test was gettting defined and the second was
being ignored.  Therefore, the Tpetra version of the test MueLu_Stratimikos
was never getting run and the Epetra version of the test MueLu_Redirection was
never getting run.

I renamed them with postfixes '_Tpetra' and '_Epetra' to allow both of the
tests to get defined and run.

I will try to allow enable these tests to see if they pass PR testing.  If
not, I will disable the test that was not getting run in the next commit.