Change method to find secondary scripts
slurm and shell runs execute the passed
script in place, but lsf copies it to a
temporary location. Since we were looking
for the merge and test python scripts in
the same directory as the driver.sh script,
this totally failed. I am now testing for
the .git directory when finding the REPO_ROOT
and then using relative paths from that to
find the python scripts.