MueLu: Add muelu_add_serial_and_mpi_test() (#8173)
This restores 1-proc tests in a non-MPI build that got removed by the changes
in PR #8115 (see #8110) and it also adds these 1-proc tests to the MPI build
as well.

With this change, the MueLu test suite for an MPI build should be a superset
of the tests run in a non-MPI build.

However, I see several 1-proc tests that are only defined to run in an MPI
build (with 'COMM mpi') but should likely also be run in a non-MPI build as
well.